requirements.txt
Flask
spotipy
python-dotenv
requests
Flask-Cors
